Hook rocket up into notion-server
